updated player wall/ground check (read notes)
may change the system again later, for now landing on top of a tree resets your jumps again.  i want to fix this, but probably going to have to give the trees their own layer to resolve.  had to make this change in order to prevent the player from being "in air" when standing on the top pixels of a wall.